Support per-user limits on coupons.
This restricts users from placing multiple orders with the same discount.
This is a commonly used feature in Magento for sites that have a max $ limit
on their coupons. Users bypass this by placing multiple orders. Magento
however doesn't force you to be logged in during the time of ordering,
so it's easily bypassed that system.