[WebBundle] Rework the common address "show" template
I don't believe this commit has much impact, nor is it particularly
important, but more of a "why not?" consideration.  So if there are no
strong oppositions, why not?

Changes:

  1. Remove <address> element.  Most opinionated google results will claim
the <address> element shouldn't be used for mailing addresses, so I've
changed this to use a schema.org microdata format.  I'm far from savvy
in this area but it does seem like schema.org is the trending microdata
dictionary if one should be adopted at all.

  2. Moved phone number to the bottom, on the assumption this is more
common across the world.  It's definitely the common format here in USA.

  3. Use country ISO code, with full country name noted on the side.  The
motivation here is that most workers managing customer information won't
have all the ISO codes memorized, but displaying the ISO code is still
useful.

- Before: http://i.imgur.com/CkwGHUy.png
- After: http://i.imgur.com/HffrXXO.png