Allow the ArchetypeUpdateListener to be used with any subject.
The `ArchetypeUpdateListener` is only compatible with Products in
Sylius, but this change allows new instances to be created that are
compatible with any subject that implements the
`ArchetypeSubjectInterface`.

Decouples from Sylius Product, allowing for it to be moved the
ArchetypeBundle.

The service `sylius.listener.archetype_update` will remain registered in
the ProductBundle as this service specifically updates Sylius Products.