[Metadata] Fix modifying metadata by serializing it early
Metadata cannot be stored as `object` ORM type, because objects are
compared (===) by reference, not by value, what means that changes
would not be persisted.

The only working way I have found is to store metadata as `text`
type and serialize it on MetadataContainer::setMetadata() method
call. However, after every metadata modification we have to use that
method once again, if not - changes won't be persisted.