Array accessible and countable Service Registry.
The idea behind this is to make it almost effortless to replace arrays
of services with a service registry. The implementations of SPL's
ArrayAccess are called from the implementations of the
RegistryInterface, the only difference being that the ArrayAccess
implementations do not throw exceptions as to align the behavior with
array usage in PHP.

The countable interface allows checking how many items are in the
registry without having to get all of them.