Adding a reference to all entities in the override doc
It's not practical to have to search for where the entity has been defined in the first place. And especially the first time you override sylius_user.resources.shop.user, it's not an obvious path. The same is for ChannelPricings.
So this could be really helpful. I'll try to keep it up to date with my further overridings.