Fix huge queries issue
In case you have a large amount of data (i.e. 100k products), the autocomplete query fetches all matching results which brokes down the whole web server. This simple fix limits the results to 25 as in most cases the end user does not need more than this amount of suggested hits. 
Could be improved by adding the limit parameter to the method signature.