Hotfix NotStandard SQL
Fixes SQL aliases in HAVING clauses so it is compatible with postgreSql and follow a more standart implementation.
With Doctrine, one of its principles is that you are not dependant on a DBSM. It could be great too make standard requests compatible with more systems.
I also propose a more easily maintainable and readable version of this provider.