Fix CheckoutResolver
Sometimes it could happen that the checkout resolver patter matches a route that does not have the `_sylius` parameters.
In such cases the following exception is thrown:
```
TypeError: Noticed exception 'TypeError' with message 'Return value of Sylius\Bundle\CoreBundle\Checkout\CheckoutResolver::getRequestedGraph() must be of the type string, null returned'
```
This change fix that.