[API] Set createdByGuest to carts created via api
The PickupCartHandler wasn't aware of the recently added createdByGuest
flag. It uses the findLatestNotEmptyCartByChannelAndCustomer method of
the cart repository, which is part of the CoreBundle, which excludes
carts created by guests. In contrast to the ShopBasedCartContext,
provided by the CoreBundle, it doesn't set the createdByGuest flag, so
all carts, created via api, are defined as created by guests. Hence, the
PickupCartHandler created a new cart every time the endpoint was called.