Widen ramsey/uuid version range
This commit allows Sylius to use either version 3 or version 4 of the
UUID package. Limiting Sylius to just 3.9.x causes deprecation notices
because it doesn't implement a return type on JsonSerializable, whereas
this is fixed in 4.x.

If downstream consumers rely on behavior specific to 3.x, they can
continue using that. However, consumers who wish to upgrade to 4.x can
do so without breaking Sylius.