Bug fix: cart completion between UI and API
Problem: you were not able to start a cart on the UI and continue on the
API. The problem was the tokenValue not generated at creation of the
cart.

This commit adds the tokenValue insertion at creation time using a
factory to enforce it at a global level.

A lot of code is related to a new feature in the behat test suite: it
was not design to alternate between API and UI easily (not that hard
though).